services/requirements/base.txt
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
dynamodb-json==1.3
dynamodbgeo==0.0.3
pynamodb==5.0.3
pynamodb-attributes==0.3.0
requests[socks]
elasticache-auto-discovery==1.0.0
redis==3.5.3
psycopg2-binary